# docs/swagger.yaml
basePath: /api
definitions:
  httpErrors.ErrResponse:
    properties:
      msg:
        description: application-level error message
        example: not found user
        type: string
      status:
        description: http response status code
        example: 404
        type: integer
      statusText:
        description: user-level status message
        example: not_found
        type: string
    type: object
...
  presenter.ItemCreate:
    properties:
      description:
        example: item description
        type: string
      title:
        example: item title
        type: string
    required:
    - description
    - title
    type: object
  presenter.ItemResponse:
    properties:
      description:
        type: string
      id:
        type: string
      owner_id:
        type: string
      title:
        type: string
    type: object
  presenter.ItemUpdate:
    properties:
      description:
        example: item description
        type: string
      title:
        example: item title
        type: string
    type: object
  responses.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/httpErrors.ErrResponse'
      is_success:
        example: false
        type: boolean
    type: object
...
  responses.SuccessResponse-array_presenter_ItemResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/presenter.ItemResponse'
        type: array
      is_success:
        example: true
        type: boolean
    type: object
...
  responses.SuccessResponse-presenter_ItemResponse:
    properties:
      data:
        $ref: '#/definitions/presenter.ItemResponse'
      is_success:
        example: true
        type: boolean
    type: object
info:
  contact: {}
  description: API for ICMON project
  title: Go Rest API
  version: "1.0"
paths:
...
  /item:
    get:
      consumes:
      - application/json
      description: Retrieve items.
      parameters:
      - description: limit
        format: limit
        in: query
        name: limit
        type: integer
      - description: offset
        format: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse-array_presenter_ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Read Items
      tags:
      - items
    post:
      consumes:
      - application/json
      description: Create new item.
      parameters:
      - description: Add item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/presenter.ItemCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse-presenter_ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Create Item
      tags:
      - items
  /item/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an item by ID.
      parameters:
      - description: Item Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse-presenter_ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Delete item
      tags:
      - items
    get:
      consumes:
      - application/json
      description: Get item by ID.
      parameters:
      - description: Item Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse-presenter_ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Read item
      tags:
      - items
    put:
      consumes:
      - application/json
      description: Update an item by ID.
      parameters:
      - description: Item Id
        in: path
        name: id
        required: true
        type: string
      - description: Update item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/presenter.ItemUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse-presenter_ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Update item
      tags:
      - items
...
securityDefinitions:
  OAuth2Password:
    flow: password
    scopes:
      read: Grants read access
      write: Grants write access
    tokenUrl: /api/auth/login
    type: oauth2
swagger: "2.0"